requirements.txt
Flask-RESTful
Flask
Flask-Cors
pyodbc 
python-dotenv
PyJWT